- Introduced CameraSource factory pattern to support both Camera & Camera2 API source because older Camera API source is required for devices where Camera2 API not supported at hardware level. - Merged additional code base required for Camera2 API within original codebase (e.g. Camera2SourcePreview into CameraSourcePreview). - Added picture size preference over preview size in GraphicOverlay for Camera2 API support.
