+++ android/material-showcase/app/src/main/java/com/google/mlkit/md/CustomModelObjectDetectionActivity.kt
@@ -41,6 +41,7 @@ import com.google.mlkit.md.camera.GraphicOverlay
 import com.google.mlkit.md.camera.WorkflowModel
 import com.google.mlkit.md.camera.WorkflowModel.WorkflowState
 import com.google.mlkit.md.camera.CameraSource
+import com.google.mlkit.md.camera.CameraSourceFactory
 import com.google.mlkit.md.camera.CameraSourcePreview
 import com.google.mlkit.md.objectdetection.MultiObjectProcessor
 import com.google.mlkit.md.objectdetection.ProminentObjectProcessor
@@ -81,7 +82,7 @@ class CustomModelObjectDetectionActivity : AppCompatActivity(), OnClickListener 
         preview = findViewById(R.id.camera_preview)
         graphicOverlay = findViewById<GraphicOverlay>(R.id.camera_preview_graphic_overlay).apply {
             setOnClickListener(this@CustomModelObjectDetectionActivity)
-            cameraSource = CameraSource(this)
+            cameraSource = CameraSourceFactory.createCameraSource(this)
         }
         promptChip = findViewById(R.id.bottom_prompt_chip)
         promptChipAnimator =
@@ -160,10 +161,10 @@ class CustomModelObjectDetectionActivity : AppCompatActivity(), OnClickListener 
             R.id.flash_button -> {
                 if (flashButton?.isSelected == true) {
                     flashButton?.isSelected = false
-                    cameraSource?.updateFlashMode(Camera.Parameters.FLASH_MODE_OFF)
+                    cameraSource?.setFlashStatus(false)
                 } else {
                     flashButton?.isSelected = true
-                    cameraSource?.updateFlashMode(Camera.Parameters.FLASH_MODE_TORCH)
+                    cameraSource?.setFlashStatus(true)
                 }
             }
             R.id.settings_button -> {
