+++ android/material-showcase/app/src/main/java/com/google/mlkit/md/camera/Camera2APISource.kt
@@ -58,29 +58,23 @@ import kotlin.math.abs
  * is unable to keep up with the rate of frames generated by the camera.
  */
 
-class Camera2Source(private val graphicOverlay: GraphicOverlay) {
+class Camera2APISource(private val graphicOverlay: GraphicOverlay): CameraSource() {
 
     private val context: Context = graphicOverlay.context
 
     /** Detects, characterizes, and connects to a CameraDevice (used for all camera operations) */
     private val cameraManager: CameraManager by lazy {
-        context.getSystemService(Context.CAMERA_SERVICE) as CameraManager
+        getCameraManager(context)
     }
 
     /** [cameraId] corresponding to the provided Camera facing back property */
     private val cameraId: String by lazy {
-        cameraManager.cameraIdList.forEach {
-            val characteristics = cameraManager.getCameraCharacteristics(it)
-            if (characteristics.get(CameraCharacteristics.LENS_FACING) == CAMERA_FACING_BACK){
-                return@lazy it
-            }
-        }
-        throw IOException("No Camera found matching the back facing lens $CAMERA_FACING_BACK")
+        getCameraId(context)
     }
 
     /** [CameraCharacteristics] corresponding to the provided Camera ID */
     private val characteristics: CameraCharacteristics by lazy {
-        cameraManager.getCameraCharacteristics(cameraId)
+        getCameraCharacteristics(context)
     }
 
     /** The [CameraDevice] that will be used for preview */
@@ -118,8 +112,10 @@ class Camera2Source(private val graphicOverlay: GraphicOverlay) {
     }
 
     /** [Size] that is currently in use by the [camera] */
-    internal var previewSize: Size? = null
-        private set
+    private var previewSize: Size? = null
+
+    /** [Size] that is use by the [imageReader] as a preference over [previewSize] If it exists */
+    private var pictureSize: Size? = null
 
     /** [Thread] for detecting & processing [imageReader] frames */
     private var processingThread: Thread? = null
@@ -133,46 +129,8 @@ class Camera2Source(private val graphicOverlay: GraphicOverlay) {
     /** [Mutex] to lock the CoroutineScope operations */
     private val mutex = Mutex()
 
-    /** [Frame2Processor] to process the frames received inside [processingRunnable] */
-    private var frameProcessor: Frame2Processor? = null
-
-
-    /**
-     * Opens the camera and starts sending preview frames to the underlying detector. The supplied
-     * surface holder is used for the preview so frames can be displayed to the user.
-     *
-     * @param surfaceHolder the surface holder to use for the preview frames.
-     * @throws Exception if the supplied surface holder could not be used as the preview display.
-     */
-    @Throws(Exception::class)
-    internal fun start(surfaceHolder: SurfaceHolder) {
-        runBlocking {
-            mutex.withLock {
-
-                if (camera != null) return@withLock
-
-                camera = createCamera().also {cameraDevice ->
-                    previewSize = getPreviewAndPictureSize(characteristics).preview.also { previewSize ->
-                        imageReader = ImageReader.newInstance(previewSize.width, previewSize.height, IMAGE_FORMAT, IMAGE_BUFFER_SIZE).also { imageReader ->
-                            session = createCaptureSession(cameraDevice, listOf(surfaceHolder.surface, imageReader.surface), cameraHandler).also {cameraCaptureSession ->
-                                captureRequest = cameraDevice.createCaptureRequest(CameraDevice.TEMPLATE_PREVIEW).apply {
-                                    addTarget(surfaceHolder.surface)
-                                    addTarget(imageReader.surface)
-                                    startPreview( this, imageReader, cameraCaptureSession)
-                                }
-                            }
-                        }
-                    }
-                    processingThread = Thread(processingRunnable).apply {
-                        processingRunnable.setActive(true)
-                        start()
-                    }
-                    relativeOrientation.observeForever(orientationObserver)
-                }
-
-            }
-        }
-    }
+    /** [FrameProcessor] to process the frames received inside [processingRunnable] */
+    private var frameProcessor: FrameProcessor? = null
 
     /**
      * Start the camera preview on the provided surface and process images through image reader buffer
@@ -192,9 +150,15 @@ class Camera2Source(private val graphicOverlay: GraphicOverlay) {
 
         //Setup listener for receiving the preview frames for processing
         imageReader.setOnImageAvailableListener({
-            it.acquireNextImage()?.let {image ->
-                val rotation = relativeOrientation.value ?: 0
-                processingRunnable.setNextFrame(image, rotation)
+            try {
+                it.acquireNextImage()?.let {image ->
+                    val rotation = relativeOrientation.value ?: 0
+                    processingRunnable.setNextFrame(image, rotation)
+                }
+            }
+            catch (e: IllegalStateException){
+                e.printStackTrace()
+                Log.e(TAG, "${e.message} At acquire next image")
             }
         }, imageReaderHandler)
 
@@ -214,70 +178,7 @@ class Camera2Source(private val graphicOverlay: GraphicOverlay) {
         session.setRepeatingRequest(captureRequest.build(), null, cameraHandler)
     }
 
-    /**
-     * Closes the camera and stops sending frames to the underlying frame detector.
-     *
-     *
-     * This camera source may be restarted again by calling [.start].
-     *
-     *
-     * Call [.release] instead to completely shut down this camera source and release the
-     * resources of the underlying detector.
-     */
-    @Throws(Exception::class)
-    internal fun stop() {
-        runBlocking {
-            mutex.withLock {
-                Log.d(TAG, "Stop is called")
-                processingRunnable.setActive(false)
-                processingThread?.let {
-                    try {
-                        // Waits for the thread to complete to ensure that we can't have multiple threads executing
-                        // at the same time (i.e., which would happen if we called start too quickly after stop).
-                        it.join()
-                    } catch (e: InterruptedException) {
-                        Log.e(TAG, "Frame processing thread interrupted on stop.")
-                    }
-                    processingThread = null
-                }
-
-                // Remove the reference image reader buffer & orientation change observer, since it will no longer be in use.
-                imageReader?.let {
-                    it.setOnImageAvailableListener(null, null)
-                    imageReader = null
-                }
-
-                relativeOrientation.removeObserver(orientationObserver)
-
-                camera?.let {
-                    it.close()
-                    camera = null
-                }
-            }
-        }
-
-    }
-
-    /** Stops the camera and releases the resources of the camera and underlying detector.  */
-    fun release() {
-        graphicOverlay.clear()
-        synchronized(processorLock) {
-            stop()
-            frameProcessor?.stop()
-            cameraThread.quitSafely()
-            imageReaderThread.quitSafely()
-        }
-    }
-
-    fun setFrameProcessor(processor: Frame2Processor) {
-        graphicOverlay.clear()
-        synchronized(processorLock) {
-            frameProcessor?.stop()
-            frameProcessor = processor
-        }
-    }
-
-    fun updateFlashMode(enabled: Boolean) {
+    private fun updateFlashMode(enabled: Boolean) {
         val flashAvailable = characteristics.get(CameraCharacteristics.FLASH_INFO_AVAILABLE) as Boolean
         if(flashAvailable){
             session?.let {session ->
@@ -357,7 +258,7 @@ class Camera2Source(private val graphicOverlay: GraphicOverlay) {
      * @throws Exception if cannot find a suitable size.
      */
     @Throws(Exception::class)
-    private fun getPreviewAndPictureSize(characteristics: CameraCharacteristics): CameraSizePair  {
+    private fun getPreviewAndPictureSize(cameraSource: CameraSource): CameraSizePair  {
 
         // Gives priority to the preview size specified by the user if exists.
         val sizePair: CameraSizePair = PreferenceUtils.getUserSpecifiedPreviewSize(context) ?: run {
@@ -369,7 +270,7 @@ class Camera2Source(private val graphicOverlay: GraphicOverlay) {
                 } else {
                     graphicOverlay.width.toFloat() / graphicOverlay.height
                 }
-            selectSizePair(characteristics, displayAspectRatioInLandscape)
+            selectSizePair(cameraSource, displayAspectRatioInLandscape)
         } ?: throw IOException("Could not find suitable preview size.")
 
         sizePair.preview.let {
@@ -386,6 +287,106 @@ class Camera2Source(private val graphicOverlay: GraphicOverlay) {
         return sizePair
     }
 
+    //Camera source overrides
+
+    override fun getSupportedPreviewSizes(): Array<Size> = characteristics.get(CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP)!!.getOutputSizes(SurfaceHolder::class.java)
+
+    override fun getSupportedPictureSizes(): Array<Size> = characteristics.get(CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP)!!.getOutputSizes(IMAGE_FORMAT)
+
+    override fun setFrameProcessor(processor: FrameProcessor) {
+        graphicOverlay.clear()
+        synchronized(processorLock) {
+            frameProcessor?.stop()
+            frameProcessor = processor
+        }
+    }
+
+    override fun setFlashStatus(status: Boolean){
+        if (status){
+            updateFlashMode(true)
+        }
+        else{
+            updateFlashMode(false)
+        }
+    }
+
+    override fun getSelectedPreviewSize() = previewSize
+
+    override fun getSelectedPictureSize() = pictureSize
+
+    override fun start(surfaceHolder: SurfaceHolder) {
+        runBlocking {
+            mutex.withLock {
+
+                if (camera != null) return@withLock
+
+                camera = createCamera().also {cameraDevice ->
+                    getPreviewAndPictureSize(this@Camera2APISource).also { sizePair ->
+                        previewSize = sizePair.preview
+                        pictureSize = sizePair.picture
+                        val imageSize = sizePair.picture ?: sizePair.preview
+                        imageReader = ImageReader.newInstance(imageSize.width, imageSize.height, IMAGE_FORMAT, IMAGE_BUFFER_SIZE).also { imageReader ->
+                            session = createCaptureSession(cameraDevice, listOf(surfaceHolder.surface, imageReader.surface), cameraHandler).also {cameraCaptureSession ->
+                                captureRequest = cameraDevice.createCaptureRequest(CameraDevice.TEMPLATE_PREVIEW).apply {
+                                    addTarget(surfaceHolder.surface)
+                                    addTarget(imageReader.surface)
+                                    startPreview( this, imageReader, cameraCaptureSession)
+                                }
+                            }
+                        }
+                    }
+                    processingThread = Thread(processingRunnable).apply {
+                        processingRunnable.setActive(true)
+                        start()
+                    }
+                    relativeOrientation.observeForever(orientationObserver)
+                }
+
+            }
+        }
+    }
+
+    override fun stop() {
+        runBlocking {
+            mutex.withLock {
+                Log.d(TAG, "Stop is called")
+                processingRunnable.setActive(false)
+                processingThread?.let {
+                    try {
+                        // Waits for the thread to complete to ensure that we can't have multiple threads executing
+                        // at the same time (i.e., which would happen if we called start too quickly after stop).
+                        it.join()
+                    } catch (e: InterruptedException) {
+                        Log.e(TAG, "Frame processing thread interrupted on stop.")
+                    }
+                    processingThread = null
+                }
+
+                // Remove the reference image reader buffer & orientation change observer, since it will no longer be in use.
+                imageReader?.let {
+                    it.setOnImageAvailableListener(null, null)
+                    imageReader = null
+                }
+
+                relativeOrientation.removeObserver(orientationObserver)
+
+                camera?.let {
+                    it.close()
+                    camera = null
+                }
+            }
+        }
+    }
+
+    override fun release() {
+        graphicOverlay.clear()
+        synchronized(processorLock) {
+            stop()
+            frameProcessor?.stop()
+            cameraThread.quitSafely()
+            imageReaderThread.quitSafely()
+        }
+    }
 
     /**
      * This runnable controls access to the underlying receiver, calling it to process frames when
@@ -513,7 +514,7 @@ class Camera2Source(private val graphicOverlay: GraphicOverlay) {
         const val CAMERA_FACING_BACK = CameraCharacteristics.LENS_FACING_BACK
         const val IMAGE_FORMAT = ImageFormat.YUV_420_888
 
-        private const val TAG = "CameraSource"
+        private const val TAG = "Camera2APISource"
 
         /** Maximum number of images that will be held in the reader's buffer */
         private const val IMAGE_BUFFER_SIZE: Int = 3
@@ -522,6 +523,21 @@ class Camera2Source(private val graphicOverlay: GraphicOverlay) {
         private const val MAX_CAMERA_PREVIEW_WIDTH = 1300
         private const val DEFAULT_REQUESTED_CAMERA_PREVIEW_WIDTH = 640
         private const val DEFAULT_REQUESTED_CAMERA_PREVIEW_HEIGHT = 360
+
+        private fun getCameraManager(context: Context) = context.getSystemService(Context.CAMERA_SERVICE) as CameraManager
+
+        private fun getCameraId(context: Context): String {
+            val cameraManager = getCameraManager(context)
+            cameraManager.cameraIdList.forEach {
+                val characteristics = cameraManager.getCameraCharacteristics(it)
+                if (characteristics.get(CameraCharacteristics.LENS_FACING) == CAMERA_FACING_BACK){
+                    return it
+                }
+            }
+            throw IOException("No Camera found matching the back facing lens $CAMERA_FACING_BACK")
+        }
+
+        fun getCameraCharacteristics(context: Context) = getCameraManager(context).getCameraCharacteristics(getCameraId(context))
 
         /**
          * Selects the most suitable preview and picture size, given the display aspect ratio in landscape
@@ -542,11 +558,11 @@ class Camera2Source(private val graphicOverlay: GraphicOverlay) {
          * ratio. On some hardware, if you would only set the preview size, you will get a distorted
          * image.
          *
-         * @param characteristics the selected camera characteristics to select a preview size from
+         * @param cameraSource the selected camera source to select a preview size from
          * @return the selected preview and picture size pair
          */
-        private fun selectSizePair(characteristics: CameraCharacteristics, displayAspectRatioInLandscape: Float): CameraSizePair? {
-            val validPreviewSizes = Utils.generateValidPreviewSizeList(characteristics)
+        private fun selectSizePair(cameraSource: CameraSource, displayAspectRatioInLandscape: Float): CameraSizePair? {
+            val validPreviewSizes = Utils.generateValidPreviewSizeList(cameraSource)
 
             var selectedPair: CameraSizePair? = null
             // Picks the preview size that has closest aspect ratio to display view.
