+++ android/material-showcase/app/src/main/java/com/google/mlkit/md/camera/FrameProcessor.kt
@@ -16,13 +16,19 @@
 
 package com.google.mlkit.md.camera
 
+import com.google.android.odml.image.MlImage
 import java.nio.ByteBuffer
 
 /** An interface to process the input camera frame and perform detection on it.  */
 interface FrameProcessor {
 
     /** Processes the input frame with the underlying detector.  */
+    @Deprecated("Keeping it only to support Camera API frame processing")
     fun process(data: ByteBuffer, frameMetadata: FrameMetadata, graphicOverlay: GraphicOverlay)
+
+    /** Processes the input frame with the underlying detector.
+     * @return true if holding [MlImage] for processing otherwise return false */
+    fun process(image: MlImage, graphicOverlay: GraphicOverlay): Boolean
 
     /** Stops the underlying detector and release resources.  */
     fun stop()
