+++ android/material-showcase/app/src/main/res/layout/activity_live_barcode.xml
@@ -5,14 +5,14 @@
     android:layout_height="match_parent"
     android:keepScreenOn="true">
 
-  <com.google.mlkit.md.camera.Camera2SourcePreview
+  <com.google.mlkit.md.camera.CameraSourcePreview
       android:id="@+id/camera_preview"
       android:layout_width="match_parent"
       android:layout_height="match_parent">
 
     <include layout="@layout/camera_preview_overlay"/>
 
-  </com.google.mlkit.md.camera.Camera2SourcePreview>
+  </com.google.mlkit.md.camera.CameraSourcePreview>
 
   <include
       layout="@layout/top_action_bar_in_live_camera"
